python-agent/pyproject.toml
[project]
name = "task-agent"
version = "0.1.0"
description = "A2A agent for task decomposition and execution"
requires-python = ">=3.10"
dependencies = [
    # A2A SDK and core dependencies (with HTTP server support)
    "a2a-sdk[http-server]>=0.3.0",

    # HTTP client for API requests
    "httpx>=0.28.1",

    # Data validation and serialization
    "pydantic>=2.11.4",

    # ASGI server for running the agent
    "uvicorn>=0.32.1",

    # CLI framework (if needed for future enhancements)
    "click>=8.1.8",

    # Additional useful packages for development
    "python-dotenv>=1.0.0",

    # OpenAI client for AI model integration
    "openai>=1.58.0",

    # Pydantic AI for agent framework
    "pydantic-ai>=0.0.14",

    # Agents framework for market analysis
    "agents>=0.1.0",

    # Rich console for enhanced terminal output
    "rich>=13.0.0",

    # Data analysis and manipulation
    "pandas>=2.0.0",

    # Yahoo Finance data fetching
    "yfinance>=0.2.0",

    # Stock statistics and technical indicators
    "stockstats>=0.5.0",

    # Date utilities for time calculations
    "python-dateutil>=2.8.0"
]

[project.scripts]
task-orchestrator = "task_agent.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_agent"]